[intersect] add basic simd implementation
Summary: This implementation uses SIMD instructions to identify if the entirety of a chunk in `b` is smaller than the value in `a`. If it is, we can skip ahead by 4 indexes in `b`.

This implementation doesn't seem to be faster, looking at the benchmarks. I'm not sure if this is because it's specialized for a small `a` and a big `b`, which isn't exactly what the benchmarks do, or if it's because i suck :|

Future options: Instead of doing lots of extra work to keep accesses to `b` aligned, we could just load things up directly. This would avoid the weird swizzling

Test Plan: cargo test